Bigass commit, with some seriously dirty code in there, but its a prototype.
This commit does a bunch of things. Basically, the salient interaction
between the android app and the platform is now in place.

The platform was refactored with stub implementations for order_pizza.
add_user is functional, but has no value checking so its not robust

The UI flow in the android app is functional for add_user and order_pizza.
This includes remembering the userId across sessions.